Update for awslocal build

+++ 3rdparty/build_openssl.sh
@@ -15,7 +15,7 @@ echo "-------------------------"
 root=$PWD
 tarfile=$pkg-$version.tar.gz
 sudo dnf install -y make gcc perl-core pcre-devel wget zlib-devel zlib-static
-if [[ ! -e $tarfile ]]; then
+if [[ ! -e $tarfile && ! -e $pkg-$version ]]; then
   wget https://github.com/openssl/openssl/releases/download/$pkg-$version/$tarfile -O $tarfile
   wget https://github.com/openssl/openssl/releases/download/$pkg-$version/$tarfile.sha256 -O $tarfile.sha256
 fi
@@ -26,12 +26,12 @@ cd $pkg-$version
 # Build & install
 #mkdir -p build
 #cd build
-./config no-shared --prefix=$install_dir --openssldir=$install_dir/ssl -- zlib
+./config no-shared --prefix=$install_dir --openssldir=$install_dir/ssl --with-zlib-lib=/usr/lib64/libz.a
 make -j$(nproc)
 make install_sw
 
 cd $root
-#rm ./$tarfile
-#rm ./$tarfile.sha256
+rm ./$tarfile
+rm ./$tarfile.sha256
 #rm -rf ./$pkg-$version
 echo "Done"
